refactor: 프로퍼티 설정으로 jdbc 연결 코드 중복 제거
