+++ src/main/java/chess/controller/IngameController.java
@@ -1,10 +1,10 @@
-package chess;
+package chess.controller;
 
 import chess.domain.ChessGame;
 import chess.domain.GameResult;
 import chess.domain.piece.Color;
 import chess.domain.position.Square;
-import chess.service.DBService;
+import chess.service.ChessService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -14,10 +14,10 @@ import java.util.List;
 
 @Controller
 public class IngameController {
-    private final DBService DBService;
+    private final ChessService chessService;
 
-    public IngameController() {
-        this.DBService = new DBService();
+    public IngameController(final ChessService chessService) {
+        this.chessService = chessService;
     }
 
     @GetMapping(value = "/ingame", params = "gameID")
@@ -37,7 +37,7 @@ public class IngameController {
 
     @PostMapping("/ingame/{gameID}")
     public String movePiece(@PathVariable String gameID, @RequestBody String movement, Model model) {
-        ChessGame chessGame = DBService.loadSavedChessGame(gameID, DBService.getTurn(gameID));
+        ChessGame chessGame = chessService.loadSavedChessGame(gameID, chessService.getTurn(gameID));
         List<String> movements = Arrays.asList(movement.split("&"));
 
         String source = getPosition(movements.get(0));
@@ -45,8 +45,8 @@ public class IngameController {
 
         try {
             chessGame.move(new Square(source), new Square(target));
-            DBService.movePiece(gameID, source, target);
-            DBService.updateTurn(gameID, chessGame);
+            chessService.movePiece(gameID, source, target);
+            chessService.updateTurn(gameID, chessGame);
             model.addAllAttributes(chessGame.getEmojis());
             model.addAttribute("msg","누가 이기나 보자구~!");
         } catch (IllegalArgumentException e) {
@@ -54,7 +54,7 @@ public class IngameController {
             model.addAttribute("msg", e.getMessage());
         }
         model.addAttribute("gameID", gameID);
-        GameResult gameResult = DBService.getGameResult(gameID);
+        GameResult gameResult = chessService.getGameResult(gameID);
         model.addAttribute("whiteScore", gameResult.calculateScore(Color.WHITE));
         model.addAttribute("blackScore", gameResult.calculateScore(Color.BLACK));
