+++ src/main/java/chess/db/ChessGameDao.java
@@ -2,27 +2,16 @@ package chess.db;
 
 import chess.domain.ChessGame;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
+@Repository
 public class ChessGameDao {
-    private final static String DRIVER = "com.mysql.jdbc.Driver";
-    private static final String URL = "jdbc:mysql://localhost:3306/chess";
-    private static final String USER = "user";
-    private static final String PASSWORD = "password";
-    private org.springframework.jdbc.datasource.DriverManagerDataSource dataSource;
     private JdbcTemplate jdbcTemplate;
 
-    public ChessGameDao() {
-        dataSource = new org.springframework.jdbc.datasource.DriverManagerDataSource(URL, USER, PASSWORD);
-        dataSource.setDriverClassName(DRIVER);
-        jdbcTemplate = new JdbcTemplate(dataSource);
+    public ChessGameDao(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
     }
 
     public void save(String gameID, ChessGame chessGame) {
