+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.h2.console.enabled=true
 
 spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.datasource.url=
-spring.datasource.username=
-spring.datasource.password=
+spring.datasource.url=jdbc:mysql://localhost:3306/chess
+spring.datasource.username=user
+spring.datasource.password=password
